Changed manapoints from int to double to keep conversion errors from happening

+++ Classes/Character.cs
@@ -15,7 +15,7 @@ namespace DH4.Classes
         public double CurrentHealthPoints{get;set;}
         public PlayerClassTypes PlayerClass{get;set;}
         public int PlayerExpPoints{get;set;}
-        public int PlayerManaAttackPoints{get; set;}
+        public double PlayerManaAttackPoints{get; set;}
 
     }
 }
